Error handling updates

+++ phangsPipeline/casaCubeRoutines.py
@@ -34,22 +34,24 @@ def copy_dropdeg(
     overwriting and importing from FITS.
     """
 
+    this_stub = 'COPY_DROPDEG: '
+
     if os.path.isdir(outfile):
         if not overwrite:
             if not quiet:
-                print("Output file already present and overwrite false. Returning.")
+                print(this_stub+"Output exists and overwrite set to false - "+outfile)
             return(False)
         os.system('rm -rf '+outfile)
 
     used_temp_outfile = False
     if (infile[-4:] == 'FITS') and os.path.isfile(infile):
         if not quiet:
-            print("Importing from FITS.")
+            print(this_stub+"Importing from FITS.")
         temp_outfile = outfile+'.temp'
         if os.path.isdir(temp_outfile):
             if not overwrite:
                 if not quiet:
-                    print("Temporary output file already present and overwrite false. Returning.")
+                    print(this_stub+"Temp file exists and overwrite set to false - "+temp_outfile)
                 return(False)
             os.system('rm -rf '+temp_outfile)
         
@@ -76,61 +78,81 @@ def primary_beam_correct(
     pbfile=None, 
     outfile=None, 
     cutoff=0.25, 
-    overwrite=False):
+    overwrite=False,
+    quiet=False):
     """
     Construct a primary-beam corrected image.
     """
+    
+    this_stub = 'PRIMARY_BEAM_CORRECT: '
 
     if infile is None or pbfile is None or outfile is None:
-        print("Missing required input.")
-        return
+        if not quiet:
+            print(this_stub+"Missing required input.")
+        return(False)
 
     if os.path.isdir(infile) == False:
-        print("Input file missing - "+infile)
-        return
+        if not quiet:
+            print(this_stub+"Input file missing - "+infile)
+        return(False)
 
     if os.path.isdir(pbfile) == False:
-        print("Primary beam file missing - "+pbfile)
-        return
+        if not quiet:
+            print(this_stub+"Primary beam file missing - "+pbfile)
+        return(False)
 
-    if overwrite:
-        os.system('rm -rf '+outfile)
+    if os.path.isfile(outfile) or os.path.isdir(outfile):
+        if overwrite:
+            os.system('rm -rf '+outfile)
+        else:            
+            if not quiet:
+                print(this_stub+"Output exists and overwrite set to false - "+outfile)
+            return(False)
 
     casa.impbcor(imagename=infile, pbimage=pbfile, outfile=outfile, cutoff=cutoff)
 
-    return()
+    return(True)
     
 def convolve_to_round_beam(
     infile=None, 
     outfile=None, 
     force_beam=None, 
-    overwrite=False):
+    overwrite=False,
+    quiet=False):
     """
     Convolve supplied image to have a round beam. Optionally, force
     that beam to some size, else it figures out the beam.
     """
     
+    this_stub = 'CONVOLVE_TO_ROUND_BEAM: '
+
     if infile is None or outfile is None:
-        print("Missing required input.")
-        return
+        if not quiet:
+            print(this_stub+"Missing required input.")
+        return(None)
 
     if os.path.isdir(infile) == False:
-        print("Input file missing - "+infile)
-        return    
+        if not quiet:
+            print(this_stub+"Input file missing - "+infile)
+        return(None)
 
     if force_beam is None:
         hdr = casa.imhead(infile)
 
         if (hdr['axisunits'][0] != 'rad'):
-            print("ERROR: Based on CASA experience. I expected units of radians.")
-            print("I did not find this. Returning. Adjust code or investigate file "+infile)
-            return
+            if not quiet:
+                print(this_stub+"ERROR: Based on CASA experience. I expected units of radians.")
+                print(this_stub+"I did not find this. Returning.")
+                print(this_stub+"Adjust code or investigate file "+infile)
+            return(None)
         pixel_as = abs(hdr['incr'][0]/np.pi*180.0*3600.)
 
         if (hdr['restoringbeam']['major']['unit'] != 'arcsec'):
-            print("ERROR: Based on CASA experience. I expected units of arcseconds for the beam.")
-            print("I did not find this. Returning. Adjust code or investigate file "+infile)
-            return    
+            if not quiet:
+                print(this_stub+"ERROR: Based on CASA experience. I expected units of arcseconds for the beam.")
+                print(this_stub+"I did not find this. Returning.")
+                print(this_stub+"Adjust code or investigate file "+infile)
+            return(None)
         bmaj = hdr['restoringbeam']['major']['value']    
         target_bmaj = np.sqrt((bmaj)**2+(2.0*pixel_as)**2)
     else:
@@ -149,11 +171,14 @@ def convolve_to_round_beam(
 
 def calc_jytok(
     hdr=None,
-    infile=None):
+    infile=None,
+    quiet=False):
     """
     Calculate the Jy/beam -> Kelvin conversion. Accepts a header
     already read using imhead or a file name that it will parse.
     """
+
+    this_stub = 'CALC_JYTOK: '
 
     c = 2.99792458e10
     h = 6.6260755e-27
@@ -161,13 +186,15 @@ def calc_jytok(
 
     if hdr is None:
         if infile is None:
-            print("No header and no infile. Returning.")
+            if not quiet:                
+                print(this_stub+"No header and no infile. Returning.")
             return(None)
         hdr = casa.imhead(target_file, mode='list')
 
     if hdr['cunit3'] != 'Hz':
-        print("I expected frequency as the third axis but did not find it.")
-        print("Returning.")
+        if not quiet:
+            print(this_stub+"I expected frequency as the third axis but did not find it.")
+            print(this_stub+"Returning.")
         return(None)
     
     crpix3 = hdr['crpix3']
@@ -179,9 +206,10 @@ def calc_jytok(
     
     bmaj_unit = hdr['beammajor']['unit']
     if bmaj_unit != 'arcsec':
-        print("Beam unit is not arcsec, which I expected. Returning.")
-        print("Unit instead is "+bmaj_unit)
-        return    
+        if not quiet:
+            print(this_stub+"Beam unit is not arcsec, which I expected. Returning.")
+            print(this_stub+"Unit instead is "+bmaj_unit)
+        return(None)
     bmaj_as = hdr['beammajor']['value']
     bmin_as = hdr['beamminor']['value']
     bmaj_sr = bmaj_as/3600.*np.pi/180.
@@ -196,26 +224,32 @@ def convert_jytok(
     infile=None, 
     outfile=None, 
     overwrite=False, 
-    inplace=False):
+    inplace=False,
+    quiet=False):
     """
     Convert a cube from Jy/beam to K.
     """
 
+    this_stub = 'CONVERT_JYTOK: '
+
     if infile is None or (outfile is None and inplace==False):
-        print("Missing required input.")
-        return
+        if not quiet:
+            print(this_stub+"Missing required input.")
+        return(False)
     
     if os.path.isdir(infile) == False:
-        print("Input file not found: "+infile)
-        return
+        if not quiet:
+            print(this_stub+"Input file not found: "+infile)
+        return(False)
     
     if inplace == False:
-        if overwrite:
-            os.system('rm -rf '+outfile)
-        
-        if os.path.isdir(outfile):
-            print("Output file already present: "+outfile)
-            return
+        if os.path.isdir(outfile) or os.path.isfile(outfile):
+            if overwrite:
+                os.system('rm -rf '+outfile)
+            else:
+                if not quiet:
+                    print(this_stub+"Output file already present: "+outfile)
+                return(False)
 
         os.system('cp -r '+infile+' '+outfile)
         target_file = outfile
@@ -225,8 +259,10 @@ def convert_jytok(
     hdr = casa.imhead(target_file, mode='list')
     unit = hdr['bunit']
     if unit != 'Jy/beam':
-        print("Unit is not Jy/beam. Returning.")
-        return
+        if not quiet:
+            print(this_stub+"Input unit is not Jy/beam for file "+target_file)
+            print(this_stub+"Instead found "+unit)
+        return(False)
     
     jytok = calc_jytok(hdr=hdr)
 
@@ -240,32 +276,38 @@ def convert_jytok(
 
     casa.imhead(target_file, mode='put', hdkey='JYTOK', hdvalue=jytok)
 
-    return()
+    return(True)
 
 def convert_ktojy(
     infile=None, 
     outfile=None, 
     overwrite=False, 
-    inplace=False):
+    inplace=False,
+    quiet=False):
     """
     Convert a cube from K to Jy/beam.
     """
 
+    this_stub = 'CONVERT_KTOJY: '
+
     if infile is None or (outfile is None and inplace==False):
-        print("Missing required input.")
-        return
+        if not quiet:
+            print(this_stub+"Missing required input.")
+        return(False)
     
     if os.path.isdir(infile) == False:
-        print("Input file not found: "+infile)
-        return
+        if not quiet:
+            print(this_stub+"Input file not found: "+infile)
+        return(False)
     
     if inplace == False:
-        if overwrite:
-            os.system('rm -rf '+outfile)
-        
-        if os.path.isdir(outfile):
-            print("Output file already present: "+outfile)
-            return
+        if os.path.isdir(outfile) or os.path.isfile(outfile):
+            if overwrite:
+                os.system('rm -rf '+outfile)
+            else:
+                if not quiet:
+                    print(this_stub+"Output file already present: "+outfile)
+                return(False)
 
         os.system('cp -r '+infile+' '+outfile)
         target_file = outfile
@@ -275,8 +317,10 @@ def convert_ktojy(
     hdr = casa.imhead(target_file, mode='list')
     unit = hdr['bunit']
     if unit != 'K':
-        print("Unit is not K. Returning.")
-        return
+        if not quiet:
+            print(this_stub+"Input unit is not K for file "+target_file)
+            print(this_stub+"Instead found "+unit)
+        return(False)
     
     jytok = calc_jytok(hdr=hdr)
 
@@ -290,14 +334,15 @@ def convert_ktojy(
 
     casa.imhead(target_file, mode='put', hdkey='JYTOK', hdvalue=jytok)
 
-    return()
+    return(True)
 
 def trim_cube(    
     infile=None, 
     outfile=None, 
     overwrite=False, 
     inplace=False, 
-    min_pixperbeam=3):
+    min_pixperbeam=3,
+    quiet=False):
     """
     Trim empty space from around the edge of a cube. Also rebin the
     cube to smaller size, while ensuring a minimum number of pixels
@@ -389,7 +434,8 @@ def export_and_cleanup(
     add_cards=[],
     add_history=[],
     zap_history=True,
-    roundbeam_tol=0.01):
+    roundbeam_tol=0.01,
+    quiet=False):
     """
     Export from a CASA image file to a FITS file, in the process
     cleaning up header keywords that are usually confusing or
